+++ lab_09_01_1/calculating_matrix.h
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <math.h>
+#include <inttypes.h>
 
 #define OK 0
 #define ERROR 221
@@ -15,17 +16,17 @@
 
 typedef struct
 {
-    int m;
-    int n;
-    int *p;
+    int64_t m;
+    int64_t n;
+    int64_t *p;
 } matrix;
 
-int find_max(matrix a_m, int *maxn, int *maxm);
-int calc_mean_column(matrix a_m, int col);
-int find_max_row(matrix a_m, int row);
-int multiply_matrix(matrix a_m, matrix b_m, matrix *m_m);
-int expo_matrix(matrix *a_m, int power);
-int copy_matrix(matrix old, matrix *newm);
+int64_t find_max(matrix a_m, int64_t *maxn, int64_t *maxm);
+int64_t calc_mean_column(matrix a_m, int64_t col);
+int64_t find_max_row(matrix a_m, int64_t row);
+int64_t multiply_matrix(matrix a_m, matrix b_m, matrix *m_m);
+int64_t expo_matrix(matrix *a_m, int64_t power);
+int64_t copy_matrix(matrix old, matrix *newm);
 
 #define LAB_09_FUNC_H
