+++ lab_09_01_1/resizing_matrix.c
@@ -1,9 +1,9 @@
 #include "calculating_matrix.h"
 #include "resizing_matrix.h"
 
-int create_matrix(matrix *a_m)
+int64_t create_matrix(matrix *a_m)
 {
-    a_m->p = malloc((a_m->n) * (a_m->m) * sizeof(int));
+    a_m->p = malloc((a_m->n) * (a_m->m) * sizeof(int64_t));
     if (!a_m->p)
     {
         puts("Create array error");
@@ -13,9 +13,9 @@ int create_matrix(matrix *a_m)
     return OK;
 }
 
-int resize_matrix(matrix *a_m, int fin)
+int64_t resize_matrix(matrix *a_m, int64_t fin)
 {
-    int *pnew = (int*)realloc(a_m->p, sizeof(pnew) * fin);
+    int64_t *pnew = (int64_t*)realloc(a_m->p, sizeof(pnew) * fin);
     if (pnew)
     {
         a_m->p = pnew;
@@ -28,11 +28,11 @@ int resize_matrix(matrix *a_m, int fin)
     return OK;
 }
 
-void delete_column(matrix *a_m, int col)
+void delete_column(matrix *a_m, int64_t col)
 {
-    for (int i = col + 1; i < a_m->n * a_m->m; i += a_m->m)
+    for (int64_t i = col + 1; i < a_m->n * a_m->m; i += a_m->m)
     {
-        for (int *j = ((a_m->p) + i); j < MAT_END; j++)
+        for (int64_t *j = ((a_m->p) + i); j < MAT_END; j++)
         {
             *(j - 1) = *j;
         }
@@ -41,12 +41,12 @@ void delete_column(matrix *a_m, int col)
     (a_m->m)--;
 }
 
-void delete_row(matrix *a_m, int row)
+void delete_row(matrix *a_m, int64_t row)
 {
-    int i = row * a_m->m + 1;
-    for (int n = 0; n < a_m->m; n++)
+    int64_t i = row * a_m->m + 1;
+    for (int64_t n = 0; n < a_m->m; n++)
     {
-        for (int *j = ((a_m->p) + i); j < MAT_END; j++)
+        for (int64_t *j = ((a_m->p) + i); j < MAT_END; j++)
         {
             *(j - 1) = *j;
         }
@@ -54,12 +54,12 @@ void delete_row(matrix *a_m, int row)
     (a_m->n)--;
 }
 
-int normalize_matrix(matrix *a_m)
+int64_t normalize_matrix(matrix *a_m)
 {
-    int maxn = 0;
-    int maxm = 0;
+    int64_t maxn = 0;
+    int64_t maxm = 0;
     find_max(*a_m, &maxn, &maxm);
-    int fin = 0;
+    int64_t fin = 0;
  
     while (a_m->n != a_m->m)
     {
@@ -80,16 +80,16 @@ int normalize_matrix(matrix *a_m)
     return OK;
 }
 
-int add_column(matrix *a_m)
+int64_t add_column(matrix *a_m)
 {
-    int newfin = a_m->m * a_m->n + a_m->n;
-    int n = 0;
+    int64_t newfin = a_m->m * a_m->n + a_m->n;
+    int64_t n = 0;
     if (resize_matrix(a_m, newfin) != OK)
         return ERROR;
     a_m->m ++;
-    for (int *i = a_m->p + a_m->m - 1; i < MAT_END; i += a_m->m)
+    for (int64_t *i = a_m->p + a_m->m - 1; i < MAT_END; i += a_m->m)
     {
-        for (int *c = MAT_END; c > i; c--)
+        for (int64_t *c = MAT_END; c > i; c--)
         {
             *c = *(c - 1);
         }
@@ -99,16 +99,16 @@ int add_column(matrix *a_m)
     return OK;
 }
 
-int add_row(matrix *a_m)
+int64_t add_row(matrix *a_m)
 {
-    int newf = a_m->m * a_m->n + a_m->m;
+    int64_t newf = a_m->m * a_m->n + a_m->m;
     if (resize_matrix(a_m, newf) != OK)
         return ERROR;
     
     a_m->n ++;
-    int n = 0;
+    int64_t n = 0;
 
-    for (int *j = MAT_END - a_m->m; j < MAT_END; j++)
+    for (int64_t *j = MAT_END - a_m->m; j < MAT_END; j++)
     {
         *j = calc_mean_column(*a_m, n);
         n++;
