+++ lab_09_01_1/resizing_matrix.h
@@ -2,11 +2,11 @@
 
 #ifndef LAB_09_RESIZING_MATRIX_H
 
-int normalize_matrix(matrix *a_m);
-int create_matrix(matrix *a_m);
-int add_column(matrix *a_m);
-int add_row(matrix *a_m);
-int resize_matrix(matrix *a_m, int fin);
+int64_t normalize_matrix(matrix *a_m);
+int64_t create_matrix(matrix *a_m);
+int64_t add_column(matrix *a_m);
+int64_t add_row(matrix *a_m);
+int64_t resize_matrix(matrix *a_m, int64_t fin);
 void equate_matrix(matrix *a_m, matrix *b_m);
 
 #define LAB_09_RESIZING_MATRIX_H
