+++ lab_09_01_1/test.c
@@ -2,7 +2,7 @@
 #include "resizing_matrix.h"
 #include "user_interface.h"
 
-void fill_square_matrix(matrix *t, int x1, int x2, int x3, int x4)
+void fill_square_matrix(matrix *t, int64_t x1, int64_t x2, int64_t x3, int64_t x4)
 {
     *t->p = x1;
     *(t->p + 1) = x2;
@@ -10,7 +10,7 @@ void fill_square_matrix(matrix *t, int x1, int x2, int x3, int x4)
     *(t->p + 3) = x4;
 }
 
-void fill_2x3_matrix(matrix *t, int x1, int x2, int x3, int x4, int x5, int x6)
+void fill_2x3_matrix(matrix *t, int64_t x1, int64_t x2, int64_t x3, int64_t x4, int64_t x5, int64_t x6)
 {
     *t->p = x1;
     *(t->p + 1) = x2;
@@ -20,7 +20,7 @@ void fill_2x3_matrix(matrix *t, int x1, int x2, int x3, int x4, int x5, int x6)
     *(t->p + 5) = x6;
 }
 
-void fill_3x3_matrix(matrix *t, int x1, int x2, int x3, int x4, int x5, int x6, int x7, int x8, int x9)
+void fill_3x3_matrix(matrix *t, int64_t x1, int64_t x2, int64_t x3, int64_t x4, int64_t x5, int64_t x6, int64_t x7, int64_t x8, int64_t x9)
 {
     *t->p = x1;
     *(t->p + 1) = x2;
@@ -33,11 +33,11 @@ void fill_3x3_matrix(matrix *t, int x1, int x2, int x3, int x4, int x5, int x6, 
     *(t->p + 8) = x9;
 }
 
-int compare_matrices(matrix a, matrix b)
+int64_t compare_matrices(matrix a, matrix b)
 {
     if (a.n != b.n || a.m != b.m)
         return FAILED;
-    for (int i = 0; i < a.m * a.n; i++)
+    for (int64_t i = 0; i < a.m * a.n; i++)
     {
         if (*(a.p + i) != *(b.p + i))
         {
@@ -47,42 +47,42 @@ int compare_matrices(matrix a, matrix b)
     return PASSED;
 }
 
-int test_normalize_ord_col(matrix *t, matrix *tres)
+int64_t test_normalize_ord_col(matrix *t, matrix *tres)
 {
     t->m = 3;
     resize_matrix(t, t->m * t->n);
     fill_square_matrix(tres, 1, 2, 4, 5);
     fill_2x3_matrix(t, 1, 2, 3, 4, 5, 6);
     normalize_matrix(t);
-    int log = compare_matrices(*t, *tres);
+    int64_t log = compare_matrices(*t, *tres);
     t->m = 2;
     resize_matrix(t, t->m * t->n);
     return log;
 }
 
-int test_normalize_ord_row(matrix *t, matrix *tres)
+int64_t test_normalize_ord_row(matrix *t, matrix *tres)
 {
     t->n = 3;
     resize_matrix(t, t->m * t->n);
     fill_square_matrix(tres, 1, 2, 3, 4);
     fill_2x3_matrix(t, 1, 2, 3, 4, 5, 6);
     normalize_matrix(t);
-    int log = compare_matrices(*t, *tres);
+    int64_t log = compare_matrices(*t, *tres);
     t->n = 2;
     resize_matrix(t, t->m * t->n);
     return log;
 }
 
-int test_normalize_same(matrix *t, matrix *tres)
+int64_t test_normalize_same(matrix *t, matrix *tres)
 {
     fill_square_matrix(tres, 1, 2, 3, 4);
     fill_square_matrix(t, 1, 2, 3, 4);
     normalize_matrix(t);
-    int log = compare_matrices(*t, *tres);
+    int64_t log = compare_matrices(*t, *tres);
     return log;
 }
 
-int test_expo_zero(matrix *t, matrix *tres)
+int64_t test_expo_zero(matrix *t, matrix *tres)
 {
     fill_square_matrix(tres, 1, 0, 0, 1);
     fill_square_matrix(t, 1, 2, 3, 4);
@@ -90,7 +90,7 @@ int test_expo_zero(matrix *t, matrix *tres)
     return compare_matrices(*t, *tres);
 }
 
-int test_expo_one(matrix *t, matrix *tres)
+int64_t test_expo_one(matrix *t, matrix *tres)
 {
     fill_square_matrix(tres, 1, 2, 3, 4);
     fill_square_matrix(t, 1, 2, 3, 4);
@@ -98,7 +98,7 @@ int test_expo_one(matrix *t, matrix *tres)
     return compare_matrices(*t, *tres);
 }
 
-int test_expo_more(matrix *t, matrix *tres)
+int64_t test_expo_more(matrix *t, matrix *tres)
 {
     fill_square_matrix(tres, 7, 10, 15, 22);
     fill_square_matrix(t, 1, 2, 3, 4);
@@ -106,7 +106,7 @@ int test_expo_more(matrix *t, matrix *tres)
     return compare_matrices(*t, *tres);
 }
 
-int test_equate_same(matrix *t, matrix *tres)
+int64_t test_equate_same(matrix *t, matrix *tres)
 {
     fill_square_matrix(t, 1, 2, 3, 4);
     fill_square_matrix(tres, 1, 2, 3, 4);
@@ -114,7 +114,7 @@ int test_equate_same(matrix *t, matrix *tres)
     return compare_matrices(*t, *tres);
 }
 
-int test_equate_ord(matrix *t, matrix *tres)
+int64_t test_equate_ord(matrix *t, matrix *tres)
 {
     t->n = 3;
     resize_matrix(t, t->m * t->n);
@@ -126,14 +126,14 @@ int test_equate_ord(matrix *t, matrix *tres)
     fill_2x3_matrix(t, 1, 2, 3, 4, 5, 6);
     fill_2x3_matrix(&sup, 1, 2, 3, 4, 2, 3);
     equate_matrix(t, tres);
-    int log = compare_matrices(sup, *tres);
+    int64_t log = compare_matrices(sup, *tres);
     t->n = 2;
     resize_matrix(t, t->m * t->n);
     free(sup.p);
     return log;
 }
 
-int test_equate_new(matrix *t, matrix *tres)
+int64_t test_equate_new(matrix *t, matrix *tres)
 {
     t->m = 3;
     resize_matrix(t, t->m * t->n);
@@ -145,7 +145,7 @@ int test_equate_new(matrix *t, matrix *tres)
     fill_2x3_matrix(t, 1, 2, 3, 4, 5, 6);
     fill_3x3_matrix(&sup, 1, 2, 2, 3, 4, 4, 2, 3, 3);
     equate_matrix(t, tres);
-    int log = compare_matrices(sup, *tres);
+    int64_t log = compare_matrices(sup, *tres);
     t->m = 2;
     resize_matrix(t, t->m * t->n);
     free(sup.p);
@@ -156,7 +156,7 @@ int main()
 {
     matrix t;
     matrix tres;
-    int c = 0;
+    int64_t c = 0;
 
     t.n = 2;
     t.m = 2;
@@ -176,7 +176,7 @@ int main()
     c += test_equate_ord(&t, &tres);
     c += test_equate_new(&t, &tres);
 
-    printf("c = %d out of 9", c);
+    printf("c = %lld out of 9", c);
 
     free(t.p);
     free(tres.p);
